Adding hexHighlight plugin as a git submodule
This will be useful for highlighting hex codes of the form #00FF00 when
trying to edit/create vim colorschemes. Note that the plugin doesn't
work as advertised (doesn't highlight #0F0), so may be buggy in other
ways. So if we don't find it that useful, we may decide to just remove
it.